refactor(memory): decompose SpectorMemoryFactory into subsystem builders
Split the ~970-line SpectorMemoryFactory god-class into an orchestrator plus
single-responsibility subsystem builders. Pure, behavior-preserving extraction:
same objects, same wiring, same dependency order, same disk/in-memory branches,
and the same #443 open-all-partitions frozen registry construction.

New package-private helpers (all in com.spectrayan.spector.memory):
- MigrationPathResolver          load-priority resolution (getNewerPath)
- CognitiveCortexBuilder         path/quantizer/namespace/partition discovery + tier stores + router
- MemoryIndexBuilder             MemoryIndex load + #443 colocated provenance + active-seq
- BiologicalSubsystemsBuilder    surprise/flashbulb/valence/coactivation/... + reflect daemon
- CognitiveGraphBuilder          hebbian/temporal/entity/hyper graphs + facade
- RetrievalIndexBuilder          BM25 (+ text store) / SPLADE / ColBERT
- CognitiveIngestionTargetBuilder ingestion target + salience wiring + active-seq seed
- PartitionManagerBuilder        frozen-partition registry + PartitionManager + roll callback + text resolver
- RecallPipelineBuilder          semantic strategy + HNSW rebuild + history + pipeline + listeners
- DaemonSupervisorBuilder        checkpoint daemon + supervisor
- MemoryWalRecovery              WAL replay / index rebuild

SpectorMemoryFactory.assemble() is now a ~120-line orchestrator that invokes the
builders in the exact prior order (cortex -> index -> WAL -> biological -> graphs
-> retrieval -> ingestion target -> partition manager -> WAL recovery -> recall
pipeline -> derived components -> daemons). File shrinks 970 -> 234 lines. No
public API change; SubsystemBundle and assemble() signatures are unchanged.

Tests: mvn test -pl memory/spector-memory -am -> 1188 pass / 0 fail / 18 skip
(identical to the pre-change baseline).

Refs #437
